feat: unoptimize report query
