+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/model/EnvActionModel.kt
@@ -12,6 +12,7 @@ import fr.gouv.cacem.monitorenv.domain.entities.mission.envAction.ActionTypeEnum
 import fr.gouv.cacem.monitorenv.domain.entities.mission.envAction.EnvActionControlPlanEntity
 import fr.gouv.cacem.monitorenv.domain.entities.mission.envAction.EnvActionEntity
 import fr.gouv.cacem.monitorenv.domain.mappers.EnvActionMapper
+import fr.gouv.cacem.monitorenv.infrastructure.database.model.reportings.ReportingModel
 import io.hypersistence.utils.hibernate.type.json.JsonBinaryType
 import jakarta.persistence.CascadeType
 import jakarta.persistence.Column
@@ -127,16 +128,16 @@ class EnvActionModel(
                     controlPlanSubThemes
                         ?.filter { subTheme ->
                             it.id.themeId ==
-                                subTheme.controlPlanSubTheme
-                                    ?.controlPlanTheme
-                                    ?.id
+                                    subTheme.controlPlanSubTheme
+                                        ?.controlPlanTheme
+                                        ?.id
                         }
                         ?.map { it.id.subthemeId },
                     tagIds =
                     controlPlanTags
                         ?.filter { tag ->
                             it.id.themeId ==
-                                tag.controlPlanTag?.controlPlanTheme?.id
+                                    tag.controlPlanTag?.controlPlanTheme?.id
                         }
                         ?.map { it.id.tagId },
                 )
